Make stack trace filtering work when concatenating assets
If Q is concatenated together with the rest of your Javascript and
served to the browser in a single file, it will filter stack traces
too aggressively, because it considers everything in the same file as
internal.

This change causes it to detect the start and end line numbers for its
own implementation so it can filter correctly even if there's other
code delivered in the same file.